requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.27,<1.0
httpx>=0.27,<1.0
websockets>=16.0,<17.0
cryptography>=42.0,<47.0
argon2-cffi>=23.1,<26.0
pydantic>=2.5,<3.0
